Build fixes: external fdk-aac and missing PMT
I was building under MSVC and I needed to build fdk-aac with different
build rules. I added a set of variables to specify includes/libs for
fdk-aac and left the exsiting external project in-tact.
Usually this would be a more formal find package for cmake, but this
handles my needs and should continue to build as-is for existing
scenarios.

Also, gr-PMT is a missing required component, I got linker errors without
it. Its probably automatically being linked in on other systems. Added
PMT to the list of components.